fix: convert server components to client components for static export compatibility
