--- app/auth/AuthPageClient.tsx
@@ -1,148 +1,0 @@
-'use client'
-
-import { useState } from 'react'
-import { createSupabaseBrowser } from '@/lib/supabase/client'
-
-interface AuthPageClientProps {
-  returnTo?: string
-  error?: string
-}
-
-export default function AuthPageClient({ returnTo, error }: AuthPageClientProps) {
-  const [isLoading, setIsLoading] = useState(false)
-  const [email, setEmail] = useState('')
-  const [message, setMessage] = useState(error || '')
-  const supabase = createSupabaseBrowser()
-  
-  const finalReturnTo = returnTo || '/'
-
-  const handleEmailSignIn = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsLoading(true)
-    setMessage('')
-
-    try {
-      const { error } = await supabase.auth.signInWithOtp({
-        email,
-        options: {
-          emailRedirectTo: `${window.location.origin}/auth/callback?returnTo=${encodeURIComponent(finalReturnTo)}`
-        }
-      })
-
-      if (error) {
-        setMessage(`Error: ${error.message}`)
-      } else {
-        setMessage('Check your email for the sign-in link!')
-      }
-    } catch (error) {
-      setMessage('An unexpected error occurred')
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const handleOAuthSignIn = async (provider: 'google' | 'github') => {
-    setIsLoading(true)
-    setMessage('')
-
-    try {
-      const { error } = await supabase.auth.signInWithOAuth({
-        provider,
-        options: {
-          redirectTo: `${window.location.origin}/auth/callback?returnTo=${encodeURIComponent(finalReturnTo)}`
-        }
-      })
-
-      if (error) {
-        setMessage(`Error: ${error.message}`)
-        setIsLoading(false)
-      }
-    } catch (error) {
-      setMessage('An unexpected error occurred')
-      setIsLoading(false)
-    }
-  }
-
-  return (
-    <div className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
-      <div className="max-w-md w-full space-y-8">
-        <div>
-          <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-            Sign in to your account
-          </h2>
-          <p className="mt-2 text-center text-sm text-gray-600">
-            Create an account or sign in to publish sales and manage favorites
-          </p>
-        </div>
-        
-        <div className="mt-8 space-y-6">
-          {/* OAuth Providers */}
-          <div className="space-y-3">
-            <button
-              onClick={() => handleOAuthSignIn('google')}
-              disabled={isLoading}
-              className="group relative w-full flex justify-center py-2 px-4 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50"
-            >
-              Continue with Google
-            </button>
-            
-            <button
-              onClick={() => handleOAuthSignIn('github')}
-              disabled={isLoading}
-              className="group relative w-full flex justify-center py-2 px-4 border border-gray-300 text-sm font-medium rounded-md text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50"
-            >
-              Continue with GitHub
-            </button>
-          </div>
-
-          <div className="relative">
-            <div className="absolute inset-0 flex items-center">
-              <div className="w-full border-t border-gray-300" />
-            </div>
-            <div className="relative flex justify-center text-sm">
-              <span className="px-2 bg-gray-50 text-gray-500">Or continue with email</span>
-            </div>
-          </div>
-
-          {/* Email Form */}
-          <form className="space-y-6" onSubmit={handleEmailSignIn}>
-            <div>
-              <label htmlFor="email" className="sr-only">
-                Email address
-              </label>
-              <input
-                id="email"
-                name="email"
-                type="email"
-                autoComplete="email"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                className="appearance-none rounded-md relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 focus:z-10 sm:text-sm"
-                placeholder="Email address"
-              />
-            </div>
-
-            <div>
-              <button
-                type="submit"
-                disabled={isLoading}
-                className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50"
-              >
-                {isLoading ? 'Sending...' : 'Send magic link'}
-              </button>
-            </div>
-          </form>
-
-          {message && (
-            <div className={`text-sm text-center ${
-              message.includes('Error') ? 'text-red-600' : 'text-green-600'
-            }`}>
-              {message}
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
-  )
-}
